requirements.txt
python-telegram-bot==13.12
python_dotenv==0.20.0
redis==4.3.4
requests==2.28.1
